requirements.txt
Django==5.2.12
django-ninja==1.5.3
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.9.0
django-extensions==4.1
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.7.0
beautifulsoup4==4.12.2
lxml==6.0.2
python-dateutil==2.8.2
requests==2.31.0
